chore(ios): upgrade Apollo iOS to 2.x, fix Swift 6 SchemaMetadata warning
Bump apollo-ios to 2.x (from 1.25.5) and enable the codegen
`markTypesNonisolated` option so generated types are marked `nonisolated`,
clearing the Swift 6 concurrency warning on `SchemaMetadata.configuration`.

Rewrite the networking layer against Apollo 2.x's async interceptor model:
- FirebaseTokenInterceptor is now an HTTPInterceptor injecting the Bearer
  token on the URLRequest.
- GraphQLLoggingInterceptor is now a GraphQLInterceptor logging errors in the
  result stream's map.
- AuthenticatedInterceptorProvider conforms to InterceptorProvider directly
  (DefaultInterceptorProvider is final in 2.x) and re-lists the defaults.
- GraphQLHelpers uses the new async fetch/perform returning GraphQLResponse,
  with `.networkOnly` keeping the cache disabled.

Convert Int call sites to Int32 where 2.x maps the GraphQL Int scalar for
operation inputs (limit/offset/row/col/rows/cols).

Generated GraphQL code is gitignored and regenerated locally via the 2.x CLI.

+++ ios/Vinarium/Features/Cellar/CellarAPI.swift
@@ -24,7 +24,7 @@ enum CellarAPI {
         let data = try await GraphQLHelpers.fetch(
             GraphQLClient.shared.apollo,
             query: VinariumGraphQL.CellarBottlesQuery(
-                limit: .some(limit),
+                limit: .some(Int32(limit)),
                 after: GraphQLHelpers.graphQLNullable(after)
             )
         )
@@ -69,7 +69,7 @@ enum CellarAPI {
     static func getHistory(limit: Int, offset: Int) async throws -> HistoryPage {
         let data = try await GraphQLHelpers.fetch(
             GraphQLClient.shared.apollo,
-            query: VinariumGraphQL.CellarHistoryQuery(limit: .some(limit), offset: .some(offset))
+            query: VinariumGraphQL.CellarHistoryQuery(limit: .some(Int32(limit)), offset: .some(Int32(offset)))
         )
         let events = data.journalEvents.items.map { e in
             HistoryEvent(
@@ -90,7 +90,7 @@ enum CellarAPI {
             GraphQLClient.shared.apollo,
             query: VinariumGraphQL.CellarGridInfoQuery()
         )
-        return CellarGridInfo(rows: data.cellarInfo.rows, cols: data.cellarInfo.cols)
+        return CellarGridInfo(rows: Int(data.cellarInfo.rows), cols: Int(data.cellarInfo.cols))
     }
 
     static func suggest() async throws -> CellarSuggestion {
@@ -108,7 +108,7 @@ enum CellarAPI {
         let rowIndex = rowIndexFromLabel(row)
         _ = try await GraphQLHelpers.perform(
             GraphQLClient.shared.apollo,
-            mutation: VinariumGraphQL.PlaceBottleMutation(beverageId: wineId, row: rowIndex, col: col)
+            mutation: VinariumGraphQL.PlaceBottleMutation(beverageId: wineId, row: Int32(rowIndex), col: Int32(col))
         )
     }
 
@@ -116,7 +116,7 @@ enum CellarAPI {
         let rowIndex = rowIndexFromLabel(row)
         _ = try await GraphQLHelpers.perform(
             GraphQLClient.shared.apollo,
-            mutation: VinariumGraphQL.MoveBottleMutation(beverageId: wineId, row: rowIndex, col: col)
+            mutation: VinariumGraphQL.MoveBottleMutation(beverageId: wineId, row: Int32(rowIndex), col: Int32(col))
         )
     }
